admin/requirements.txt
Django==4.0.4
djangorestframework==3.13.1
django-mysql==3.9
django-cors-headers==3.5.0
pika==1.1.0
python-dotenv==0.20.0
mysqlclient==2.0.1